+++ src/hooks/mutations/useDeleteTodo.tsx
@@ -3,7 +3,7 @@ import { DeleteTodoParamType, MutateOptionsType } from '@/hooks/mutations/types'
 import { todoKeys } from '@/hooks/queries';
 import { InfiniteQueryData } from '@/hooks/queries/types';
 import { ApiResponse, ToDoDetail } from '@/types';
-import { optimisticUpdateMutateHandler } from '@/utils/optimisticUpdateMutateHandler';
+import { optimisticUpdateMutateHandler } from '@/utils';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 const useDeleteTodo = (
