+++ src/utils/handleKeyDown.ts
@@ -1,9 +1,7 @@
 import { KeyboardEvent } from 'react';
 
-const handleKeyDown = (e: KeyboardEvent<HTMLElement>, callback: VoidFunction) => {
+export const handleKeyDown = (e: KeyboardEvent<HTMLElement>, callback: VoidFunction) => {
   if (e.key === 'Enter') {
     callback();
   }
 };
-
-export default handleKeyDown;
